project/dwd-2/requirements.txt
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
boto3==1.35.14
botocore==1.35.14
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==0.29.36
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.1.0
fonttools==4.53.1
fqdn==1.5.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
idna==3.8
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.4.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
monotonic==1.6
msgpack==1.0.8
neptune==1.11.1
nest-asyncio==1.6.0
numpy==1.26.2
nvidia-nccl-cu12==2.22.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.10.1
seaborn==0.13.2
simplejson==3.19.3
six==1.16.0
sklearn-contrib-py-earth==1.13+3.gf4a579e.dirty
smmap==5.0.1
stack-data==0.6.3
swagger-spec-validator==3.0.4
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
websocket-client==1.8.0
xgboost==2.1.1
